Enums

+++ koltinx-serialization-scale/src/main/java/io/novasama/substrate_sdk_android/koltinx_serialization_scale/binary/annotations/Enums.kt
@@ -4,4 +4,4 @@ import kotlinx.serialization.SerialInfo
 
 @SerialInfo
 @Target(AnnotationTarget.PROPERTY, AnnotationTarget.FIELD, AnnotationTarget.CLASS)
-annotation class EnumIndex(val index: Byte)
+annotation class EnumIndex(val index: Int)
